requirements.txt
alembic==1.5.4
amqp==5.0.5
appdirs==1.4.4
arrow==0.17.0
billiard==3.6.3.0
black==20.8b1
blinker==1.4
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
flake8==3.8.4
Flask==1.1.2
Flask-DebugToolbar==0.11.0
Flask-Migrate==2.6.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
kombu==5.0.2
Mako==1.1.4
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.1
prompt-toolkit==3.0.15
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2021.1
redis==3.5.3
regex==2020.11.13
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.23
toml==0.10.2
twilio==6.51.1
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
Werkzeug==1.0.1
WTForms==2.3.3
